% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decathlon.R
\docType{data}
\name{decathlon}
\alias{decathlon}
\title{Performance in decathlon (data)}
\format{The data used here refer to athletes' performance during two sporting events.
\describe{
 A data frame with 41 rows and 13 columns: the first ten columns corresponds to the performance of the athletes for the 10 events of the decathlon. The columns 11 and 12 correspond respectively to the rank and the points obtained. The last column is a categorical variable corresponding to the sporting event (2004 Olympic Game or 2004 Decastar)
}}
\source{
Department of statistics and computer science, Agrocampus Rennes
}
\usage{
data(decathlon)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
