% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConvexHullPCA.R
\name{plotConvexHullPCA}
\alias{plotConvexHullPCA}
\title{Plot convex hull when performing PCA}
\usage{
plotConvexHullPCA(pc, axes, zfactor, convex = T, varnames, indnames,
  biplot = T, varlab, indlab, alpha = 0.4)
}
\arguments{
\item{pc}{A FactomineR object obtained after using PCA function}

\item{axes}{It's a vector of length 2 containing the rank of the components that we want to plot}

\item{zfactor}{It's a vector from the original data, it' the sumplementary factor variable used in the PCA}

\item{convex}{It's a boolean function indicates where or not we plot the Hull Convex}

\item{varnames}{It's a vector containing the names of the active variables}

\item{indnames}{It's a vector containing the names of the individuals, i.e; the rownames}

\item{biplot}{It's a boolean vector indicates whether or not we make a biplot or a simple individual plot}

\item{varlab}{It's vector with values c("text","point","arrow")}

\item{indlab}{It's vector with values c("text","point"}

\item{alpha}{A numeric parameter with values from 0 to 1 and it gives the transparency of the plot}
}
\value{
Individual factorial map and Biplot from a PCA
}
\description{
This function produces the  plot of the individual factorial map and the biplot from a PCA object obtained after perfoming a Principal Components Analysis with FactoMineR package
}
\examples{
data(decathlon)
require(FactoMineR)
pc<-PCA(decathlon,scale.unit = T,quanti.sup = 11:12,quali.sup = 13,graph = F)
plotConvexHullPCA(pc = pc,axes = c(1,2),zfactor = decathlon$Competition,convex = T,varnames = colnames(decathlon)[1:10],indnames = rownames(decathlon),varlab = c("text","arrow"),indlab = c("text"))
}
\author{
Dhafer Malouche
}
